+++ koreaguide-api/src/main/java/com/example/koreaguide/ifs/CrudInterface.java
@@ -13,10 +13,10 @@ import javax.validation.Valid;
 public interface CrudInterface<Req,Res> {
     Header<Res> create(Header<Req> request);
 
-    Header<Res> read(Long id);
+    Header<Res> read(Integer id);
 
-    Header<Res> update(Long id, Header<Req> request);
+    Header<Res> update(Integer id, Header<Req> request);
 
-    Header delete(Long id);
+    Header delete(Integer id);
 
 }
