+++ koreaguide-api/src/main/java/com/example/koreaguide/utils/JwtUtil.java
@@ -4,7 +4,6 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
-import org.springframework.stereotype.Component;
 
 import java.security.Key;
 
@@ -13,7 +12,7 @@ public class JwtUtil {
     public JwtUtil(String secret){
         this.key = Keys.hmacShaKeyFor(secret.getBytes());
     }
-    public String createAccessToken(Long id, String name){
+    public String createAccessToken(Integer id, String name){
         //TODO: JJWT 사용
         String secret = "12345678901234567890123456789012";
